Always initialise edgehash.  The symptoms were running off the end of linked-lists and getting caught in a loop.
An explanation given by ragge:

"... due to the nature of instruction generation when special needs are
involved it get used.  In liveness analysis temporary registers needed to
keep values between instructions are usually not entered into any live
table because of their short-lived nature (when traversing the tree it's
known which instructions it will interfere with).  But; if the instruction
have special needs or so then the variables will be kept live for a longer
time and therefore added to the hash table."

Also includes an indent fix and wrap useful debugging information in PCC_DEBUG.